ramips: mt7621: (APP-3302, PR #286) Move PWM LED from kmod to kernel builtin
The status LED is now driven by the PWM driver. When the driver is built as
a kmod module, it is loaded after the preinit state is set. This means
the diag.sh script cannot set the color and flashing pattern. Making the
driver built in means the LED can be controlled at preinit stage.

Approved-by: James Haggerty
Approved-by: Sophronia Koilpillai