ramips/mt76x8: (APP-3436, PR #301) update kernel config to exclude RAVE_SP
If you don’t have this a clean build now prompts you to select
an option for RAVE_SP (causing a failure unless you run with V=sc
and answer the option).

I'm a bit confused about why this is suddenly necessary AND why the
default make kernel_*config doesn't populate it, but it's echoing a
change that went in with the matter PR for artini.

I’m also concerned that this means we probably should be regenerating
other kernel configs (since people might take this SDK and expect it
to work for other archiectures), but the good news there is that most
of them already have the RAVE thingy.


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn