mt7621: (APP-3777, PR #341) fix halowlink/artini 2.4 LED
We dropped a commit on rebasing which forced the morse module
to load late. Now, on Artini, the morse module takes phy0 before
the 2.4 module does.

Ideally we'd solve this with some kind of hotplug magic IMO, but
the standard OpenWrt LED setup seems to just hardcode this
(I guess the bringup is close enough to deterministic).


Approved-by: Arien Judge
Approved-by: Lyall Beveridge