umdns: (APP-4174, PR #407) clear cache on interface removal
This is taken from upstream, but is more significant than you might think.
This is because if you don't clear the cache it still has references
to the freed ifaces, so it ends up crashing at some point in the future.

What's worse, when it crashes, even though umdns has respawn
it doesn't do set_config again. This means that any umdns crash
means it never recovers unless it's reconfigured, since set_config
is necessary to make it start listening.


Approved-by: Sophronia Koilpillai
Approved-by: Lyall Beveridge