kernel: (APP-4688, PR #433) minor tweaks after bluetooth config change
Without this change, the bcm bluetooth module wasn't built on non-bcm
boards but it was still included in the package so building
the ipk would fail. This change makes the bcm module be built
like the other modules.

I think the question re RAVE_SP is also related to the bluetooth
change, but I've seen it pop up way too many times to fix it again
in a specific config file (and putting a not set in is a safe operation).


Approved-by: Sophronia Koilpillai
Approved-by: Evan Benn