boards: (APP-4792, PR #448) stop including tlm firmware by default
Aside from it being a bit weird (why do we want this firmware on non
mm6108 devices?), this meant that if you didn't select a normal
firmware (e.g. mm6108/mm8108) the driver would still be included
since the tlm firmware would be there. This is confusing for
users of our fork bringing up new devices.

Approved-by: Arien Judge
Approved-by: Evan Benn