uhttpd: (APP-1523, PR #458) increase default script_timeout
The cgi-upload script (part of cgi-io) doesn't send any responses
while the upload is happening, only when it's finished.
Ideally this should be re-architected to stream 'something', but
for now, the simple option is just to bump the timeout here.

At 1Mbps, this means we should just be able to upload a 30MiB firmware
image within the timeout.


Approved-by: Sophronia Koilpillai
Approved-by: Prabha Murugesan